Add files via upload
This embedded system project aims to create a simple real-time operating system (RTOS) scheduler using the ATmega32 microcontroller. The scheduler will manage three tasks, each controlling the state of an LED (on or off) in specific time periods. This project serves as an introduction to multitasking and task scheduling in embedded systems.